requirements.txt
python-dotenv
fastapi
numpy
requests
pymongo
pyodbc
uvicorn